+++ Assets/Scripts/Player/IPlayer.cs
@@ -15,6 +15,6 @@ public interface IPlayer
     Weapon GetWeapon();
     bool AddItem(CollectableItemSO item);
     bool RemoveItem(CollectableItemSO item);
-    void SetMaxHP(float x);
+    void SetMaxHP(int x);
     void SetMaxStamina(float x);
 }
