+++ Assets/Scripts/Player/PlayerEvents.cs
@@ -3,11 +3,11 @@ using UnityEngine;
 
 public class PlayerEvents
 {
-    public event Action<float, Vector3> OnTakeDamage;
-    public void TakeDamage(float value, Vector3 source) => OnTakeDamage?.Invoke(value, source);
+    public event Action<int, Vector3> OnTakeDamage;
+    public void TakeDamage(int value, Vector3 source) => OnTakeDamage?.Invoke(value, source);
     
-    public event Action<float> OnHealthRegen;
-    public void RegenHealth(float value) => OnHealthRegen?.Invoke(value);
+    public event Action<int> OnHealthRegen;
+    public void RegenHealth(int value) => OnHealthRegen?.Invoke(value);
     
     public event Action<float> OnDreamShardsChange;
     public void DreamShardsChange(float value) => OnDreamShardsChange?.Invoke(value);
